Backend/requirements.txt
langchain
langchain-community
fastapi
pymongo
scikit-learn
ipykernel
python-dotenv
pypdf